// sitat-fe-workspace/projects/app-commons/src/lib/modules/app-commons/models/tecnico/tecnico.model.ts
export interface Tecnico {

    /**
     * CAP del RUP
     */
    cap?: string;
    /**
     * CF del RUP
     */
    cf?: string;
    /**
     * Codice Istat comune del RUP
     */
    codIstat?: string;
    /**
     * Codice del RUP
     */
    codice?: string;
    /**
     * Cognome del RUP
     */
    cognome?: string;
    /**
     * Comune del RUP
     */
    comune?: string;
    /**
     * Email del RUP
     */
    email?: string;
    /**
     * Fax del RUP
     */
    fax?: string;
    /**
     * Indirizzo del RUP
     */
    indirizzo?: string;
    /**
     * Nome del RUP
     */
    nome?: string;
    /**
     * Nominativo del RUP
     */
    nominativo?: string;
    /**
     * Numero civico del RUP
     */
    numCivico?: string;
    /**
     * Provincia del RUP
     */
    provincia?: string;
    /**
     * Codice stazione appaltante del RUP
     */
    stazioneAppaltante?: string;
    /**
     * Telefono del RUP
     */
    telefono?: string;
    
    nominativoStazioneAppaltante?: string;
}

export interface RupInsertForm {
    /**
     * CAP del RUP
     */
    cap?: string;
    /**
     * CF del RUP
     */
    cf?: string;
    /**
     * Codice Istat comune del RUP
     */
    codIstat?: string;
    /**
     * Codice del RUP
     */
    codice?: string;
    /**
     * Cognome del RUP
     */
    cognome?: string;
    /**
     * Comune del RUP
     */
    comune?: string;
    /**
     * Email del RUP
     */
    email?: string;
    /**
     * Fax del RUP
     */
    fax?: string;
    /**
     * Indirizzo del RUP
     */
    indirizzo?: string;
    /**
     * Nome del RUP
     */
    nome?: string;
    /**
     * Nominativo del RUP
     */
    nominativo?: string;
    /**
     * Numero civico del RUP
     */
    numCivico?: string;
    /**
     * Provincia del RUP
     */
    provincia?: string;
    /**
     * Codice stazione appaltante del RUP
     */
    stazioneAppaltante?: string;
    /**
     * Telefono del RUP
     */
    telefono?: string;
}